requirements.txt
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==4.0.3
django-cors-headers==3.11.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
drf-yasg==1.20.0
executing==0.8.3
idna==3.3
inflection==0.5.1
ipython==8.1.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
pyparsing==3.0.7
python-dotenv==0.19.2
pytz==2021.3
random-username==1.0.2
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
traitlets==5.1.1
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
